Further refactoring and fixed HP bar bug
